+++ CHANGELOG.md
@@ -7,3 +7,12 @@
 ## 0.1.0
 
 - Support for map, get, and post
+
+## 0.2.0
+
+- Support for multiple maps in constructor
+- Support for delete, head, patch, and put
+- Support requestHandled listener
+- Support erroredHandler
+- Add creation tests
+- Add errored tests